fix: vulnerability details for qualys port policy
