+++ jobs/pacman-awsrules/src/main/java/com/tmobile/cloud/awsrules/ec2/EC2withSeverityVulnerabilityRule.java
@@ -27,6 +27,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.google.gson.JsonArray;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
@@ -100,16 +101,15 @@ public class EC2withSeverityVulnerabilityRule extends BasePolicy {
 		if (resourceAttributes != null) {
 			instanceId = StringUtils.trim(resourceAttributes.get(PacmanRuleConstants.INSTANCEID));
 			try {
-				List<String> severityList = PacmanUtils.getSeverityVulnerabilitiesByInstanceId(instanceId,ec2WithVulnInfoForS5Url, severityVulnValue);
-				if (!severityList.isEmpty()) {
-					
+				JsonArray severityList = PacmanUtils.getSeverityVulnerabilitiesArray(instanceId,ec2WithVulnInfoForS5Url, severityVulnValue);
+				if (severityList!=null && !severityList.isEmpty()) {
+					String severityDetails = PacmanUtils.getQualysVulnerabilitiesDetails(severityList);
 					annotation = Annotation.buildAnnotation(ruleParam,Annotation.Type.ISSUE);
 					annotation.put(PacmanSdkConstants.DESCRIPTION,"An Ec2 instance with vulnerability "+ severityVulnValue + " found!!");
 					annotation.put(PacmanRuleConstants.SEVERITY, severity);
 					annotation.put(PacmanRuleConstants.CATEGORY, category);
-					
+					annotation.put("vulnerabilityDetails", severityDetails);
 					issue.put(PacmanRuleConstants.VIOLATION_REASON, "An Ec2 instance with vulnerability "+ severityVulnValue + " found!!");
-					issue.put("voilation_title", String.join(",", severityList));
 					issueList.add(issue);
 					annotation.put("issueDetails",issueList.toString());
