+++ jobs/pacman-awsrules/src/main/java/com/tmobile/cloud/awsrules/ec2/ResourceWithSeverityVulnerabilityRule.java
@@ -53,10 +53,6 @@ import com.tmobile.pacman.commons.policy.PolicyResult;
 public class ResourceWithSeverityVulnerabilityRule extends BasePolicy {
 
 	private static final Logger logger = LoggerFactory.getLogger(ResourceWithSeverityVulnerabilityRule.class);
-	public static final String NIST_VULN_DETAILS_URL = "https://nvd.nist.gov/vuln/detail/";
-
-	AwsSecretManagerUtil awsSecretManagerUtil=new AwsSecretManagerUtil();
-	CredentialProvider credentialProvider=new CredentialProvider();
 
 	/**
 	 * The method will get triggered from Rule Engine with following parameters
@@ -106,17 +102,15 @@ public class ResourceWithSeverityVulnerabilityRule extends BasePolicy {
 			String entityType = resourceAttributes.get(PacmanRuleConstants.ENTITY_TYPE);
 			instanceId = StringUtils.trim(resourceAttributes.get(PacmanRuleConstants.RESOURCE_ID));
 			try {
-				List<String> violationTitle  = PacmanUtils.getSeverityVulnerabilitiesByInstanceId(instanceId,resourceWithVulnInfoForSeverityUrl, severityVulnValue);
 				JsonArray severityList = PacmanUtils.getSeverityVulnerabilitiesArray(instanceId,resourceWithVulnInfoForSeverityUrl, severityVulnValue);
 				if (severityList!=null && !severityList.isEmpty()) {
-					String severityDetails = getQualysVulnerabilitiesDetails(severityList);
+					String severityDetails = PacmanUtils.getQualysVulnerabilitiesDetails(severityList);
 					annotation = Annotation.buildAnnotation(ruleParam,Annotation.Type.ISSUE);
 					annotation.put(PacmanSdkConstants.DESCRIPTION,entityType+" instance with vulnerability "+ severityVulnValue + " found!!");
 					annotation.put(PacmanRuleConstants.SEVERITY, severity);
 					annotation.put(PacmanRuleConstants.CATEGORY, category);
 					annotation.put("vulnerabilityDetails", severityDetails);
 					issue.put(PacmanRuleConstants.VIOLATION_REASON, entityType+" instance with vulnerability "+ severityVulnValue + " found!!");
-					issue.put("violation_title", String.join(",", violationTitle));
 					issueList.add(issue);
 					annotation.put("issueDetails",issueList.toString());
 					//annotation.put("qualysIssueDetails", String.join(",", severityDetails.toString()));
@@ -136,65 +130,6 @@ public class ResourceWithSeverityVulnerabilityRule extends BasePolicy {
 
 	public String getHelpText() {
 		return "This rule checks for an EC2 Instance or VM having severity (S5,S4,S3) vulnerability if so reports it as an issue with severity High";
-	}
-
-	private String getQualysVulnerabilitiesDetails(JsonArray vulnerabilities) {
-
-		List<QualysVulnerabilityinfo> vulnerabilityList=new ArrayList<>();
-		if(vulnerabilities!=null){
-			String vulnerabityDetailurl=getVulDetailUrl();
-			for (int i = 0; i < vulnerabilities.size(); i++) {
-				JsonObject source = vulnerabilities.get(i).getAsJsonObject().get(PacmanRuleConstants.SOURCE)
-						.getAsJsonObject();
-				QualysVulnerabilityinfo vulnerabilityinfo=new QualysVulnerabilityinfo();
-				String title=source.get("title").getAsString();
-				vulnerabilityinfo.setTitle(title);
-				String hostInstanceVulnId=source.get("hostInstanceVulnId").getAsBigDecimal().toPlainString();
-				vulnerabilityinfo.setVulnerabilityUrl(vulnerabityDetailurl+hostInstanceVulnId);
-
-				JsonObject cveList = source.get("cvelist").getAsJsonObject();
-				List<CveDetails> cveDeytailsList=new ArrayList<>();
-				if(cveList!=null){
-					populateCveList(cveList, cveDeytailsList);
-				}
-				vulnerabilityinfo.setCveList(cveDeytailsList);
-				vulnerabilityList.add(vulnerabilityinfo);
-			}
-		}
-		ObjectMapper objectMapper = new ObjectMapper();
-		try {
-			return objectMapper.writeValueAsString(vulnerabilityList);
-		} catch (JsonProcessingException e) {
-			throw new RuleExecutionFailedExeption(e.getMessage());
-		}
-
-	}
-
-	private String getVulDetailUrl() {
-		String baseAccount=System.getProperty(Constants.BASE_ACCOUNT);
-		String baseRegion=System.getProperty(Constants.BASE_REGION);
-		String baseRole=System.getProperty(Constants.BASE_ROLE);
-		String secretManagerPrefix=System.getProperty(Constants.SECRET_MANAGER_PATH);
-
-		BasicSessionCredentials credential = credentialProvider.getBaseAccountCredentials(baseAccount,baseRegion, baseRole);
-		String secretData=awsSecretManagerUtil.fetchSecret(secretManagerPrefix+"/"+baseRole+"/qualys",credential,baseRegion);
-		Map<String, String> dataMap = PacmanUtils.getJsonData(secretData);
-		String baseUri=dataMap.get("qualysApiUrl");
-		baseUri=baseUri.replace("qualysapi","qualysguard")+"/vm/#/vulndetails/";
-		logger.info("Qualys base url for vulnerability details: {} ",baseUri);
-		return baseUri;
-	}
-
-	private static void populateCveList(JsonObject cveList, List<CveDetails> cveDeytailsList) {
-		JsonArray cveArray= cveList.get("cve").getAsJsonArray();
-		if(cveArray!=null){
-			for (int j = 0; j < cveArray.size(); j++) {
-				String id = cveArray.get(j).getAsJsonObject().get(PacmanRuleConstants.ID).getAsString();
-				String url = NIST_VULN_DETAILS_URL +id;
-				CveDetails cveDetails=new CveDetails(id,url);
-				cveDeytailsList.add(cveDetails);
-			}
-		}
 	}
 
 }
